+++ lightengine-client/src/main/java/net/luxvacuos/lightengine/client/rendering/api/nanovg/shaders/Window3DShader.java
@@ -20,7 +20,8 @@
 
 package net.luxvacuos.lightengine.client.rendering.api.nanovg.shaders;
 
-import net.luxvacuos.igl.vector.Matrix4d;
+import org.joml.Matrix4f;
+
 import net.luxvacuos.lightengine.client.core.ClientVariables;
 import net.luxvacuos.lightengine.client.ecs.entities.CameraEntity;
 import net.luxvacuos.lightengine.client.rendering.api.opengl.shaders.ShaderProgram;
@@ -48,7 +49,7 @@ public class Window3DShader extends ShaderProgram {
 		super.stop();
 	}
 
-	public void loadTransformationMatrix(Matrix4d matrix) {
+	public void loadTransformationMatrix(Matrix4f matrix) {
 		transformationMatrix.loadMatrix(matrix);
 	}
 
@@ -56,7 +57,7 @@ public class Window3DShader extends ShaderProgram {
 		viewMatrix.loadMatrix(camera.getViewMatrix());
 	}
 
-	public void loadProjectionMatrix(Matrix4d projection) {
+	public void loadProjectionMatrix(Matrix4f projection) {
 		projectionMatrix.loadMatrix(projection);
 	}
