+++ lightengine-client/src/main/java/net/luxvacuos/lightengine/client/rendering/api/opengl/shaders/EntityShader.java
@@ -20,7 +20,8 @@
 
 package net.luxvacuos.lightengine.client.rendering.api.opengl.shaders;
 
-import net.luxvacuos.igl.vector.Matrix4d;
+import org.joml.Matrix4f;
+
 import net.luxvacuos.lightengine.client.core.ClientVariables;
 import net.luxvacuos.lightengine.client.ecs.entities.CameraEntity;
 import net.luxvacuos.lightengine.client.rendering.api.opengl.objects.Material;
@@ -47,7 +48,7 @@ public class EntityShader extends ShaderProgram {
 		super.storeAllUniformLocations(transformationMatrix, projectionMatrix, viewMatrix, material);
 	}
 
-	public void loadTransformationMatrix(Matrix4d matrix) {
+	public void loadTransformationMatrix(Matrix4f matrix) {
 		transformationMatrix.loadMatrix(matrix);
 	}
 
@@ -59,7 +60,7 @@ public class EntityShader extends ShaderProgram {
 		viewMatrix.loadMatrix(camera.getViewMatrix());
 	}
 
-	public void loadProjectionMatrix(Matrix4d projection) {
+	public void loadProjectionMatrix(Matrix4f projection) {
 		projectionMatrix.loadMatrix(projection);
 	}
 }
