+++ lightengine-client/src/main/java/net/luxvacuos/lightengine/client/rendering/api/opengl/shaders/IrradianceCaptureShader.java
@@ -20,7 +20,8 @@
 
 package net.luxvacuos.lightengine.client.rendering.api.opengl.shaders;
 
-import net.luxvacuos.igl.vector.Matrix4d;
+import org.joml.Matrix4f;
+
 import net.luxvacuos.lightengine.client.core.ClientVariables;
 import net.luxvacuos.lightengine.client.ecs.entities.CameraEntity;
 import net.luxvacuos.lightengine.client.rendering.api.opengl.shaders.data.Attribute;
@@ -46,7 +47,7 @@ public class IrradianceCaptureShader extends ShaderProgram {
 		viewMatrix.loadMatrix(camera.getViewMatrix());
 	}
 
-	public void loadProjectionMatrix(Matrix4d projection) {
+	public void loadProjectionMatrix(Matrix4f projection) {
 		projectionMatrix.loadMatrix(projection);
 	}
